Add re-writing confirmation

+++ scripts/fastq_modules.py
@@ -39,6 +39,12 @@ def write_filtered_fastq(sequences, output_fastq):
 
     output_file_path = os.path.join(filtered_dir, output_fastq)
 
+    if os.path.exists(output_file_path):
+        confirm = input(f"File '{output_file_path}' already exists. Do you wish to re-write it? (y/n)")
+        if confirm.lower() != 'y':
+            print("Writing canceled")
+            return
+
     with open(output_file_path, "w") as fastq_file:
         for seq_id, (sequence, quality) in sequences.items():
             fastq_file.write(f"@{seq_id}\n")
